example/requirements.txt
Django==4.2.3
psycopg==3.1.14
django-filter==2.4.0
django-crispy-forms==1.13.0
